+++ Hybrid_A_Star/reeds-shepp/reeds_shepp.cpp
@@ -3,14 +3,14 @@
 
 // The comments, variable names, etc. use the nomenclature from the Reeds & Shepp paper.
 
-const double pi = M_PI;
-const double twopi = 2. * pi;
-const double RS_EPS = 1e-6;
-const double ZERO = 10*std::numeric_limits<double>::epsilon();
+const float pi = M_PI;
+const float twopi = 2. * pi;
+const float RS_EPS = 1e-6;
+const float ZERO = 10*std::numeric_limits<float>::epsilon();
 
-inline double mod2pi(double x)
+inline float mod2pi(float x)
 {
-  double v = fmod(x, twopi);
+  float v = fmod(x, twopi);
   if (v < -pi)
       v += twopi;
   else
@@ -18,21 +18,21 @@ inline double mod2pi(double x)
           v -= twopi;
   return v;
 }
-inline void polar(double x, double y, double &r, double &theta)
+inline void polar(float x, float y, float &r, float &theta)
 {
   r = sqrt(x*x + y*y);
   theta = atan2(y, x);
 }
-inline void tauOmega(double u, double v, double xi, double eta, double phi, double &tau, double &omega)
+inline void tauOmega(float u, float v, float xi, float eta, float phi, float &tau, float &omega)
 {
-  double delta = mod2pi(u-v), A = sin(u) - sin(delta), B = cos(u) - cos(delta) - 1.;
-  double t1 = atan2(eta*A - xi*B, xi*A + eta*B), t2 = 2. * (cos(delta) - cos(v) - cos(u)) + 3;
+  float delta = mod2pi(u-v), A = sin(u) - sin(delta), B = cos(u) - cos(delta) - 1.;
+  float t1 = atan2(eta*A - xi*B, xi*A + eta*B), t2 = 2. * (cos(delta) - cos(v) - cos(u)) + 3;
   tau = (t2<0) ? mod2pi(t1+pi) : mod2pi(t1);
   omega = mod2pi(tau - u + v - phi) ;
 }
 
 // formula 8.1 in Reeds-Shepp paper
-inline bool LpSpLp(double x, double y, double phi, double &t, double &u, double &v)
+inline bool LpSpLp(float x, float y, float phi, float &t, float &u, float &v)
 {
   polar(x - sin(phi), y - 1. + cos(phi), u, t);
   if (t >= -ZERO)
@@ -49,14 +49,14 @@ inline bool LpSpLp(double x, double y, double phi, double &t, double &u, double 
   return false;
 }
 // formula 8.2
-inline bool LpSpRp(double x, double y, double phi, double &t, double &u, double &v)
+inline bool LpSpRp(float x, float y, float phi, float &t, float &u, float &v)
 {
-  double t1, u1;
+  float t1, u1;
   polar(x + sin(phi), y - 1. - cos(phi), u1, t1);
   u1 = u1*u1;
   if (u1 >= 4.)
   {
-      double theta;
+      float theta;
       u = sqrt(u1 - 4.);
       theta = atan2(2., u);
       t = mod2pi(t1 + theta);
@@ -68,9 +68,9 @@ inline bool LpSpRp(double x, double y, double phi, double &t, double &u, double 
   }
   return false;
 }
-void CSC(double x, double y, double phi, ReedsSheppStateSpace::ReedsSheppPath &path)
+void CSC(float x, float y, float phi, ReedsSheppStateSpace::ReedsSheppPath &path)
 {
-  double t, u, v, Lmin = path.length(), L;
+  float t, u, v, Lmin = path.length(), L;
   if (LpSpLp(x, y, phi, t, u, v) && Lmin > (L = fabs(t) + fabs(u) + fabs(v)))
   {
       path = ReedsSheppStateSpace::ReedsSheppPath(
@@ -118,9 +118,9 @@ void CSC(double x, double y, double phi, ReedsSheppStateSpace::ReedsSheppPath &p
           ReedsSheppStateSpace::reedsSheppPathType[13], -t, -u, -v);
 }
 // formula 8.3 / 8.4  *** TYPO IN PAPER ***
-inline bool LpRmL(double x, double y, double phi, double &t, double &u, double &v)
+inline bool LpRmL(float x, float y, float phi, float &t, float &u, float &v)
 {
-  double xi = x - sin(phi), eta = y - 1. + cos(phi), u1, theta;
+  float xi = x - sin(phi), eta = y - 1. + cos(phi), u1, theta;
   polar(xi, eta, u1, theta);
   if (u1 <= 4.)
   {
@@ -134,9 +134,9 @@ inline bool LpRmL(double x, double y, double phi, double &t, double &u, double &
   }
   return false;
 }
-void CCC(double x, double y, double phi, ReedsSheppStateSpace::ReedsSheppPath &path)
+void CCC(float x, float y, float phi, ReedsSheppStateSpace::ReedsSheppPath &path)
 {
-  double t, u, v, Lmin = path.length(), L;
+  float t, u, v, Lmin = path.length(), L;
   if (LpRmL(x, y, phi, t, u, v) && Lmin > (L = fabs(t) + fabs(u) + fabs(v)))
   {
       path = ReedsSheppStateSpace::ReedsSheppPath(
@@ -163,7 +163,7 @@ void CCC(double x, double y, double phi, ReedsSheppStateSpace::ReedsSheppPath &p
   }
 
   // backwards
-  double xb = x*cos(phi) + y*sin(phi), yb = x*sin(phi) - y*cos(phi);
+  float xb = x*cos(phi) + y*sin(phi), yb = x*sin(phi) - y*cos(phi);
   if (LpRmL(xb, yb, phi, t, u, v) && Lmin > (L = fabs(t) + fabs(u) + fabs(v)))
   {
       path = ReedsSheppStateSpace::ReedsSheppPath(
@@ -187,9 +187,9 @@ void CCC(double x, double y, double phi, ReedsSheppStateSpace::ReedsSheppPath &p
           ReedsSheppStateSpace::reedsSheppPathType[1], -v, -u, -t);
 }
 // formula 8.7
-inline bool LpRupLumRm(double x, double y, double phi, double &t, double &u, double &v)
+inline bool LpRupLumRm(float x, float y, float phi, float &t, float &u, float &v)
 {
-  double xi = x + sin(phi), eta = y - 1. - cos(phi), rho = .25 * (2. + sqrt(xi*xi + eta*eta));
+  float xi = x + sin(phi), eta = y - 1. - cos(phi), rho = .25 * (2. + sqrt(xi*xi + eta*eta));
   if (rho <= 1.)
   {
       u = acos(rho);
@@ -202,9 +202,9 @@ inline bool LpRupLumRm(double x, double y, double phi, double &t, double &u, dou
   return false;
 }
 // formula 8.8
-inline bool LpRumLumRp(double x, double y, double phi, double &t, double &u, double &v)
+inline bool LpRumLumRp(float x, float y, float phi, float &t, float &u, float &v)
 {
-  double xi = x + sin(phi), eta = y - 1. - cos(phi), rho = (20. - xi*xi - eta*eta) / 16.;
+  float xi = x + sin(phi), eta = y - 1. - cos(phi), rho = (20. - xi*xi - eta*eta) / 16.;
   if (rho>=0 && rho<=1)
   {
       u = -acos(rho);
@@ -219,9 +219,9 @@ inline bool LpRumLumRp(double x, double y, double phi, double &t, double &u, dou
   }
   return false;
 }
-void CCCC(double x, double y, double phi, ReedsSheppStateSpace::ReedsSheppPath &path)
+void CCCC(float x, float y, float phi, ReedsSheppStateSpace::ReedsSheppPath &path)
 {
-  double t, u, v, Lmin = path.length(), L;
+  float t, u, v, Lmin = path.length(), L;
   if (LpRupLumRm(x, y, phi, t, u, v) && Lmin > (L = fabs(t) + 2.*fabs(u) + fabs(v)))
   {
       path = ReedsSheppStateSpace::ReedsSheppPath(
@@ -270,13 +270,13 @@ void CCCC(double x, double y, double phi, ReedsSheppStateSpace::ReedsSheppPath &
           ReedsSheppStateSpace::reedsSheppPathType[3], -t, -u, -u, -v);
 }
 // formula 8.9
-inline bool LpRmSmLm(double x, double y, double phi, double &t, double &u, double &v)
+inline bool LpRmSmLm(float x, float y, float phi, float &t, float &u, float &v)
 {
-  double xi = x - sin(phi), eta = y - 1. + cos(phi), rho, theta;
+  float xi = x - sin(phi), eta = y - 1. + cos(phi), rho, theta;
   polar(xi, eta, rho, theta);
   if (rho >= 2.)
   {
-      double r = sqrt(rho*rho - 4.);
+      float r = sqrt(rho*rho - 4.);
       u = 2. - r;
       t = mod2pi(theta + atan2(r, -2.));
       v = mod2pi(phi - .5*pi - t);
@@ -288,9 +288,9 @@ inline bool LpRmSmLm(double x, double y, double phi, double &t, double &u, doubl
   return false;
 }
 // formula 8.10
-inline bool LpRmSmRm(double x, double y, double phi, double &t, double &u, double &v)
+inline bool LpRmSmRm(float x, float y, float phi, float &t, float &u, float &v)
 {
-  double xi = x + sin(phi), eta = y - 1. - cos(phi), rho, theta;
+  float xi = x + sin(phi), eta = y - 1. - cos(phi), rho, theta;
   polar(-eta, xi, rho, theta);
   if (rho >= 2.)
   {
@@ -304,9 +304,9 @@ inline bool LpRmSmRm(double x, double y, double phi, double &t, double &u, doubl
   }
   return false;
 }
-void CCSC(double x, double y, double phi, ReedsSheppStateSpace::ReedsSheppPath &path)
+void CCSC(float x, float y, float phi, ReedsSheppStateSpace::ReedsSheppPath &path)
 {
-  double t, u, v, Lmin = path.length() - .5*pi, L;
+  float t, u, v, Lmin = path.length() - .5*pi, L;
   if (LpRmSmLm(x, y, phi, t, u, v) && Lmin > (L = fabs(t) + fabs(u) + fabs(v)))
   {
       path = ReedsSheppStateSpace::ReedsSheppPath(
@@ -358,7 +358,7 @@ void CCSC(double x, double y, double phi, ReedsSheppStateSpace::ReedsSheppPath &
   }
 
   // backwards
-  double xb = x*cos(phi) + y*sin(phi), yb = x*sin(phi) - y*cos(phi);
+  float xb = x*cos(phi) + y*sin(phi), yb = x*sin(phi) - y*cos(phi);
   if (LpRmSmLm(xb, yb, phi, t, u, v) && Lmin > (L = fabs(t) + fabs(u) + fabs(v)))
   {
       path = ReedsSheppStateSpace::ReedsSheppPath(
@@ -407,8 +407,8 @@ void CCSC(double x, double y, double phi, ReedsSheppStateSpace::ReedsSheppPath &
              ReedsSheppStateSpace::reedsSheppPathType[11], -v, -u, .5*pi, -t);
 }
 // formula 8.11 *** TYPO IN PAPER ***
-inline bool LpRmSLmRp(double x, double y, double phi, double &t, double &u, double &v) {
-  double xi = x + sin(phi), eta = y - 1. - cos(phi), rho, theta;
+inline bool LpRmSLmRp(float x, float y, float phi, float &t, float &u, float &v) {
+  float xi = x + sin(phi), eta = y - 1. - cos(phi), rho, theta;
   polar(xi, eta, rho, theta);
   if (rho >= 2.) {
     u = 4. - sqrt(rho*rho - 4.);
@@ -425,9 +425,9 @@ inline bool LpRmSLmRp(double x, double y, double phi, double &t, double &u, doub
   return false;
 }
 
-void CCSCC(double x, double y, double phi, ReedsSheppStateSpace::ReedsSheppPath &path)
+void CCSCC(float x, float y, float phi, ReedsSheppStateSpace::ReedsSheppPath &path)
 {
-  double t, u, v, Lmin = path.length() - pi, L;
+  float t, u, v, Lmin = path.length() - pi, L;
   if (LpRmSLmRp(x, y, phi, t, u, v) && Lmin > (L = fabs(t) + fabs(u) + fabs(v)))
   {
       path = ReedsSheppStateSpace::ReedsSheppPath(
@@ -451,7 +451,7 @@ void CCSCC(double x, double y, double phi, ReedsSheppStateSpace::ReedsSheppPath 
           ReedsSheppStateSpace::reedsSheppPathType[17], -t, .5*pi, -u, .5*pi, -v);
 }
 
-ReedsSheppStateSpace::ReedsSheppPath reedsShepp(double x, double y, double phi)
+ReedsSheppStateSpace::ReedsSheppPath reedsShepp(float x, float y, float phi)
 {
   ReedsSheppStateSpace::ReedsSheppPath path;
   CSC(x, y, phi, path);
@@ -485,7 +485,7 @@ ReedsSheppStateSpace::reedsSheppPathType[18][5] = {
 };
 
 ReedsSheppStateSpace::ReedsSheppPath::ReedsSheppPath(const ReedsSheppPathSegmentType* type,
-double t, double u, double v, double w, double x)
+float t, float u, float v, float w, float x)
 : type_(type)
   {
   length_[0] = t; length_[1] = u; length_[2] = v; length_[3] = w; length_[4] = x;
@@ -493,20 +493,20 @@ double t, double u, double v, double w, double x)
 }
 
 
-double ReedsSheppStateSpace::distance(double q0[3], double q1[3])
+float ReedsSheppStateSpace::distance(float q0[3], float q1[3])
 {
   return rho_ * reedsShepp(q0, q1).length();
 }
 
-ReedsSheppStateSpace::ReedsSheppPath ReedsSheppStateSpace::reedsShepp(double q0[3], double q1[3])
+ReedsSheppStateSpace::ReedsSheppPath ReedsSheppStateSpace::reedsShepp(float q0[3], float q1[3])
 {
-  double dx = q1[0] - q0[0], dy = q1[1] - q0[1], dth = q1[2] - q0[2];
-  double c = cos(q0[2]), s = sin(q0[2]);
-  double x = c*dx + s*dy, y = -s*dx + c*dy;
+  float dx = q1[0] - q0[0], dy = q1[1] - q0[1], dth = q1[2] - q0[2];
+  float c = cos(q0[2]), s = sin(q0[2]);
+  float x = c*dx + s*dy, y = -s*dx + c*dy;
   return ::reedsShepp(x/rho_, y/rho_, dth);
 }
 
-ReedsSheppStateSpace::path_types ReedsSheppStateSpace::type(double q0[3], double q1[3]) {
+ReedsSheppStateSpace::path_types ReedsSheppStateSpace::type(float q0[3], float q1[3]) {
   path_types types;
   ReedsSheppPath path = reedsShepp(q0, q1);
   for (int i = 0; i < 5; i++) {
@@ -515,27 +515,27 @@ ReedsSheppStateSpace::path_types ReedsSheppStateSpace::type(double q0[3], double
   return types;
 }
 
-ReedsSheppStateSpace::sample_paths ReedsSheppStateSpace::sample(double q0[3], double q1[3], 
-                                          double step_size) {
+ReedsSheppStateSpace::sample_paths ReedsSheppStateSpace::sample(float q0[3], float q1[3], 
+                                          float step_size) {
   sample_paths samples;
   ReedsSheppPath path = reedsShepp(q0, q1);
-  double dist = rho_ * path.length();
+  float dist = rho_ * path.length();
 
-  for (double seg = 0.0; seg <= dist; seg += step_size){
-    vector<double> qnew = {0.0, 0.0, 0.0};
+  for (float seg = 0.0; seg <= dist; seg += step_size){
+    vector<float> qnew = {0.0, 0.0, 0.0};
     interpolate(q0, path, seg/rho_, qnew);
     samples.push_back(qnew);
   }
   return samples;
 }
 
-void ReedsSheppStateSpace::interpolate(double q0[3], ReedsSheppPath& path, 
-                                       double seg, vector<double>& s) {
+void ReedsSheppStateSpace::interpolate(float q0[3], ReedsSheppPath& path, 
+                                       float seg, vector<float>& s) {
                       
   if (seg < 0.0) seg = 0.0;
   if (seg > path.length()) seg = path.length();
 
-  double phi, v;
+  float phi, v;
 
   s[0] = s[1] = 0.0;
   s[2] = q0[2];
